requirements.txt
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
python-dotenv==1.0.1
boto3==1.34.51
sagemaker==2.212.0
jupyter==1.0.0
matplotlib==3.8.3
seaborn==0.13.2
